refactor: MetaData 추가 및 바로팟 수정 페이지 리팩토링
